sandbox/python/requirements.txt
pika==1.3.2
pytest==8.3.5
python-dotenv==1.0.1